Main loop update.

+++ TestConsoleApp/Neurone.cs
@@ -9,7 +9,7 @@ namespace TestConsoleApp
         {
         }
 
-        public double NeuroneFunction(List<float> input, List<float> weight, float bias, string function)
+        public float NeuroneFunction(List<float> input, List<float> weight, float bias, string function)
         {
             float result = 0;
             for (int i = 0; i <= input.Count; i++)
